+++ Assets/Scripts/Weapon/WeaponStats.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "WeaponStats", menuName = "Weapon/Weapon Data")]
 public class WeaponStats : ScriptableObject
 {
-    public int LifeTime;
+    public float LifeTime;
     public int Speed;
     public int Damage;
     public int Penetration;
